Move AST node ID stuff to its own module
