+++ millc.cabal
@@ -6,7 +6,7 @@ cabal-version:          >=1.10
 
 library
     hs-source-dirs:     src-lib
-    exposed-modules:    Language.Mill.AST, Language.Mill.Lex, Language.Mill.Parse
+    exposed-modules:    Language.Mill.AST, Language.Mill.AST.ID, Language.Mill.Lex, Language.Mill.Parse
     default-language:   Haskell2010
     build-depends:      base, parsec
